Extracted utils into separate files
