+++ cronjobs/batchTransactions/batchBucket.ts
@@ -1,19 +1,17 @@
 import { Psbt } from "bitcoinjs-lib";
-import { FEE, NETWORK } from "../../constants";
+import { NETWORK } from "../../constants";
+import { attemptPushToBucket } from "../../src/utils/batch/attemptPushToBucket";
 import { getExcessMiningFees } from "../../src/utils/batch/getExcessMiningFees";
 import { sha256 } from "../../src/utils/crypto/sha256";
 import { db } from "../../src/utils/db";
 import { KEYS } from "../../src/utils/db/keys";
-import { finalize, signAllInputs } from "../../src/utils/psbt";
+import { mapPSBTToDensity } from "../../src/utils/psbt/mapPSBTToDensity";
 import { isDefined } from "../../src/utils/validation";
-import { getSignerByIndex } from "../../src/wallets/getSignerByIndex";
-import { hotWallet, oldHotWallet } from "../../src/wallets/hotWallet";
 import { logger } from "./batchTransactions";
 import { finalizeBatch } from "./finalizeBatch";
 import { getServiceFees } from "./getServiceFees";
 import { getUTXOForInput } from "./getUTXOForInput";
 import { inputIsUnspent } from "./helpers/inputIsUnspent";
-import { sumPSBTInputValues } from "./helpers/sumPSBTInputValues";
 
 const BYTES_PER_INPUT = 41;
 const MAX_ACCEPTABLE_LOSS = 0.02;
@@ -53,7 +51,16 @@ export const batchBucket = async (
   const bucket: Psbt[] = [];
   for (const { psbt } of sortedPsbts) {
     // eslint-disable-next-line no-await-in-loop -- we need to wait for the result of the function
-    await attemptPushToBucket(psbt, bucket, feeRateThreshold);
+    const { wasAdded, finalFeeRate } = await attemptPushToBucket(
+      psbt,
+      bucket,
+      feeRateThreshold,
+    );
+    if (!wasAdded) {
+      logger.info([
+        `Skipping PSBT ${sha256(psbt.toBase64())} - Fee threshold is ${feeRateThreshold} but final fee rate is ${finalFeeRate}`,
+      ]);
+    }
   }
   if (bucket.length === 0) {
     return { error: "No PSBTs could be batched" };
@@ -94,47 +101,3 @@ export const batchBucket = async (
     },
   };
 };
-
-async function mapPSBTToDensity(psbt: Psbt) {
-  const psbtCopy = Psbt.fromBase64(psbt.toBase64(), { network: NETWORK });
-  const inputValues = sumPSBTInputValues(psbtCopy);
-  const index = await db.client.hGet(
-    KEYS.PSBT.PREFIX + sha256(psbtCopy.toBase64()),
-    "index",
-  );
-
-  if (isDefined(index)) {
-    signAllInputs(
-      psbtCopy,
-      getSignerByIndex(hotWallet, index, NETWORK),
-      getSignerByIndex(oldHotWallet, index, NETWORK),
-    );
-  }
-  const tx = finalize(psbtCopy);
-  const serviceFees = Math.round(inputValues * FEE);
-  const miningFees = psbtCopy.getFee() - serviceFees;
-  const feeRate = miningFees / tx.virtualSize();
-  return {
-    psbt,
-    density: serviceFees / (1 / feeRate),
-  };
-}
-
-async function attemptPushToBucket(
-  psbt: Psbt,
-  bucket: Psbt[],
-  feeRateThreshold: number,
-) {
-  const bucketWithPSBT = [...bucket, psbt];
-  const serviceFees = getServiceFees(bucketWithPSBT);
-
-  const { stagedTx } = await finalizeBatch(bucketWithPSBT, serviceFees);
-  const finalFeeRate = stagedTx.getFeeRate();
-  if (finalFeeRate >= feeRateThreshold) {
-    bucket.push(psbt);
-  } else {
-    logger.info([
-      `Skipping PSBT ${sha256(psbt.toBase64())} - Fee threshold is ${feeRateThreshold} but final fee rate is ${finalFeeRate}`,
-    ]);
-  }
-}
